updating a few more small areas
