--- aws-golang-dynamo-stream-to-es/Gopkg.toml
@@ -1,46 +1,0 @@
-# Gopkg.toml example
-#
-# Refer to https://github.com/golang/dep/blob/master/docs/Gopkg.toml.md
-# for detailed Gopkg.toml documentation.
-#
-# required = ["github.com/user/thing/cmd/thing"]
-# ignored = ["github.com/user/project/pkgX", "bitbucket.org/user/project/pkgA/pkgY"]
-#
-# [[constraint]]
-#   name = "github.com/user/project"
-#   version = "1.0.0"
-#
-# [[constraint]]
-#   name = "github.com/user/project2"
-#   branch = "dev"
-#   source = "github.com/myfork/project2"
-#
-# [[override]]
-#   name = "github.com/x/y"
-#   version = "2.4.0"
-#
-# [prune]
-#   non-go = false
-#   go-tests = true
-#   unused-packages = true
-
-
-[[constraint]]
-  name = "github.com/aws/aws-lambda-go"
-  version = "1.4.0"
-
-[[constraint]]
-  name = "github.com/aws/aws-sdk-go"
-  version = "1.15.13"
-
-[[constraint]]
-  name = "github.com/olivere/elastic"
-  version = "6.1.27"
-
-[[constraint]]
-  name = "github.com/satori/go.uuid"
-  version = "1.2.0"
-
-[prune]
-  go-tests = true
-  unused-packages = true
